Submit V1

+++ api/authors/delete.php
@@ -1,28 +1,44 @@
 <?php
-  // Headers
+  // error_reporting(E_ALL);
+  // ini_set('display_errors', 1);
+
+  // CORS
   header('Access-Control-Allow-Origin: *');
   header('Content-Type: application/json');
   header('Access-Control-Allow-Methods: DELETE');
-  header('Access-Control-Allow-Headers: Access-Control-Allow-Headers, Content-Type, Access-Control-Allow-Methods, Authorization,X-Requested-With');
+  header('Access-Control-Allow-Headers: Access-Control-Allow-Headers, Content-Type, Access-Control-Allow-Methods, Authorization, X-Requested-With');
 
-  // Include necessary files
+  // Necessary files
   include_once '../../config/Database.php';
   include_once '../../models/Author.php';
 
-// Instantiate Database object
-$database = new Database();
-$db = $database->connect();
+  // Create Database object
+  $database = new Database();
+  $db = $database->connect();
 
-// Instantiate Author object
-$author = new Author($db);
+  // Create Author object
+  $author = new Author($db);
 
-// Get author id from URL
-$author->id = isset($_GET['id']) ? $_GET['id'] : die();
+  // Get raw posted data
+  $data = json_decode(file_get_contents("php://input"));
 
-// Delete author
-if ($author->delete()) {
-    echo json_encode(array('message' => 'Author deleted'));
-} else {
-    echo json_encode(array('message' => 'Author not deleted'));
-}
+  if (!empty($data->id)) 
+  {
+      $author->id = $data->id;
+
+      // Delete author
+      if ($author->delete()) 
+      {
+          // Return ID of deleted author
+          echo json_encode(array('id' => $author->id));
+      } else 
+      {
+          echo json_encode(array('message' => 'Author not deleted'));
+      }
+  } else 
+  {
+      // ID is missing
+      echo json_encode(array('message' => 'Missing Required Parameter: id'));
+      die();
+  }
 ?>
